metadata: centralize ssa fact recording
