+++ ssa/abitype.go
@@ -26,8 +26,6 @@ import (
 
 	"github.com/goplus/llgo/ssa/abi"
 	"github.com/xgo-dev/llvm"
-
-	"github.com/goplus/llgo/internal/metadata"
 )
 
 // -----------------------------------------------------------------------------
@@ -345,71 +343,6 @@ func (b Builder) abiExtendedFields(t types.Type, name string) (fields []llvm.Val
 	return
 }
 
-func (b Builder) recordTypeChildren(parentName string, t types.Type) {
-	mb := b.Pkg.MetaBuilder
-	if mb == nil {
-		return
-	}
-	parent := mb.Symbol(parentName)
-	for _, child := range b.directTypeChildren(t) {
-		childName, _ := b.Prog.abi.TypeName(child)
-		mb.AddTypeChild(parent, mb.Symbol(childName))
-	}
-}
-
-func (b Builder) directTypeChildren(t types.Type) []types.Type {
-	switch t := types.Unalias(t).(type) {
-	case *types.Basic:
-		return nil
-	case *types.Pointer:
-		return []types.Type{abi.PublicType(t.Elem())}
-	case *types.Chan:
-		return []types.Type{abi.PublicType(t.Elem())}
-	case *types.Slice:
-		return []types.Type{abi.PublicType(t.Elem())}
-	case *types.Array:
-		elem := abi.PublicType(t.Elem())
-		return []types.Type{elem, types.NewSlice(elem)}
-	case *types.Map:
-		return []types.Type{
-			abi.PublicType(t.Key()),
-			abi.PublicType(t.Elem()),
-			// Map type descriptors reference their runtime bucket type too.
-			b.Prog.abi.MapBucket(t),
-		}
-	case *types.Signature:
-		var children []types.Type
-		children = appendTupleTypeChildren(children, t.Params())
-		children = appendTupleTypeChildren(children, t.Results())
-		return children
-	case *types.Struct:
-		children := make([]types.Type, 0, t.NumFields())
-		for i := 0; i < t.NumFields(); i++ {
-			children = append(children, abi.PublicType(t.Field(i).Type()))
-		}
-		return children
-	case *types.Interface:
-		children := make([]types.Type, 0, t.NumMethods())
-		for i := 0; i < t.NumMethods(); i++ {
-			children = append(children, funcType(b.Prog, t.Method(i).Type()))
-		}
-		return children
-	case *types.Named:
-		return b.directTypeChildren(t.Underlying())
-	}
-	return nil
-}
-
-func appendTupleTypeChildren(children []types.Type, tuple *types.Tuple) []types.Type {
-	if tuple == nil {
-		return children
-	}
-	for i := 0; i < tuple.Len(); i++ {
-		children = append(children, abi.PublicType(tuple.At(i).Type()))
-	}
-	return children
-}
-
 func (b Builder) abiUncommonPkg(t types.Type) (*types.Package, string) {
 retry:
 	switch typ := types.Unalias(t).(type) {
@@ -491,11 +424,7 @@ func (b Builder) abiUncommonMethods(t types.Type, mset *types.MethodSet) llvm.Va
 	ft := prog.rtType("Method")
 	n := mset.Len()
 	fields := make([]llvm.Value, n)
-	var slots []metadata.MethodSlot
-	typeName, _ := prog.abi.TypeName(t)
-	if b.Pkg.MetaBuilder != nil {
-		slots = make([]metadata.MethodSlot, 0, n)
-	}
+	methodInfo := b.newMethodInfoRecorder(t, n)
 	pkg, _ := b.abiUncommonPkg(t)
 	anonymous := pkg == nil
 	if anonymous {
@@ -526,21 +455,9 @@ func (b Builder) abiUncommonMethods(t types.Type, mset *types.MethodSet) llvm.Va
 		values = append(values, ifn)
 		values = append(values, tfn)
 		fields[i] = llvm.ConstNamedStruct(ft.ll, values)
-		if mb := b.Pkg.MetaBuilder; mb != nil {
-			mtypeName, _ := prog.abi.TypeName(ftyp)
-			slots = append(slots, metadata.MethodSlot{
-				Sig: metadata.MethodSig{
-					Name:  mb.Name(fullName),
-					MType: mb.Symbol(mtypeName),
-				},
-				IFn: mb.Symbol(ifn.Name()),
-				TFn: mb.Symbol(tfn.Name()),
-			})
-		}
+		methodInfo.add(obj, ftyp, ifn, tfn)
 	}
-	if mb := b.Pkg.MetaBuilder; mb != nil {
-		mb.AddMethodInfo(mb.Symbol(typeName), slots)
-	}
+	methodInfo.finish()
 	return llvm.ConstArray(ft.ll, fields)
 }
 
@@ -548,14 +465,6 @@ func (b Builder) abiUncommonMethods(t types.Type, mset *types.MethodSet) llvm.Va
 func funcType(prog Program, typ types.Type) types.Type {
 	ftyp := prog.Type(typ, InGo)
 	return ftyp.raw.Type.(*types.Struct).Field(0).Type()
-}
-
-func mthName(method *types.Func) string {
-	name := method.Name()
-	if token.IsExported(name) {
-		return name
-	}
-	return abi.FullName(method.Pkg(), name)
 }
 
 func (b Builder) abiMethodFunc(anonymous bool, mPkg *types.Package, mName string, mSig *types.Signature) (tfn llvm.Value) {
